refactor: 백엔드 오류코드 수정
