+++ Back-End/src/main/java/net/scit/backend/channel/handler/CustomWebSocketHandler.java
@@ -2,7 +2,7 @@ package net.scit.backend.channel.handler;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.scit.backend.auth.JwtTokenProvider;
+import net.scit.backend.jwt.JwtTokenProvider;
 import net.scit.backend.channel.session.WebSocketSessionManager;
 
 import org.springframework.web.socket.CloseStatus;
