+++ Back-End/src/main/java/net/scit/backend/notification/controller/NotificationController.java
@@ -2,7 +2,7 @@ package net.scit.backend.notification.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.scit.backend.auth.AuthUtil;
+import net.scit.backend.jwt.AuthUtil;
 import net.scit.backend.notification.dto.NotificationRequestDTO;
 import net.scit.backend.notification.entity.NotificationEntity;
 import net.scit.backend.notification.service.NotificationService;
@@ -24,7 +24,6 @@ public class NotificationController {
     private final NotificationService notificationService;
     private final WorkspaceMemberRepository workspaceMemberRepository;
 
-
     // ✅ SSE 구독 엔드포인트 (JWT 기반)
     @GetMapping("/subscribe")
     public SseEmitter subscribe(@RequestHeader("Authorization") String token) {
@@ -33,26 +32,29 @@ public class NotificationController {
     }
 
     /**
-     *  ✅ 특정 사용자의 읽지 않은 알림 조회 (JWT 기반)
+     * ✅ 특정 사용자의 읽지 않은 알림 조회 (JWT 기반)
+     * 
      * @param token
      * @return
      */
     @GetMapping("/unread")
-    public ResponseEntity<List<NotificationEntity>> getUnreadNotifications(@RequestHeader("Authorization") String token) {
+    public ResponseEntity<List<NotificationEntity>> getUnreadNotifications(
+            @RequestHeader("Authorization") String token) {
         String email = AuthUtil.getLoginUserId();
         List<NotificationEntity> unreadNotifications = notificationService.getUnreadNotifications(email);
         return ResponseEntity.ok(unreadNotifications);
     }
 
-
     /**
      * / ✅ 알림 개별 읽음 처리 (JWT 기반)
+     * 
      * @param token
      * @param notificationNumber
      * @return
      */
     @PostMapping("/read-single")
-    public ResponseEntity<String> markAsRead(@RequestHeader("Authorization") String token, @RequestParam Long notificationNumber) {
+    public ResponseEntity<String> markAsRead(@RequestHeader("Authorization") String token,
+            @RequestParam Long notificationNumber) {
         boolean result = notificationService.markAsRead(notificationNumber);
         return result
                 ? ResponseEntity.ok("해당 알림을 읽는 데 성공하였습니다")
@@ -60,7 +62,8 @@ public class NotificationController {
     }
 
     /**
-     *  ✅ 알림 모두 읽음 처리 (JWT 기반)
+     * ✅ 알림 모두 읽음 처리 (JWT 기반)
+     * 
      * @param token
      * @return
      */
@@ -74,9 +77,9 @@ public class NotificationController {
                 : ResponseEntity.ok("읽지 않은 알림이 없습니다.");
     }
 
-
     /**
      * ✅ 알림 삭제 (JWT 기반 + 워크스페이스 검증 추가)
+     * 
      * @param token
      * @param notificationNumber
      * @param workspaceId
@@ -84,8 +87,8 @@ public class NotificationController {
      */
     @DeleteMapping
     public ResponseEntity<String> deleteNotification(@RequestHeader("Authorization") String token,
-                                                     @RequestParam Long notificationNumber,
-                                                     @RequestParam Long workspaceId) {
+            @RequestParam Long notificationNumber,
+            @RequestParam Long workspaceId) {
         try {
             String email = AuthUtil.getLoginUserId();
 
@@ -107,9 +110,9 @@ public class NotificationController {
         }
     }
 
-
     /**
      * ✅ 새로운 알림을 수동으로 생성 (JWT 기반 + 워크스페이스 검증 추가)
+     * 
      * @param token
      * @param request
      * @return
@@ -144,8 +147,7 @@ public class NotificationController {
             String responseMessage = String.format(
                     "알림 생성에 성공하였습니다. [알림 제목: %s, 유형: %s]",
                     request.getNotificationName(),
-                    request.getNotificationType()
-            );
+                    request.getNotificationType());
 
             return ResponseEntity.ok(responseMessage);
         } catch (Exception e) {
